requirements.txt
sentence-transformers
faiss-cpu
google-generativeai
pandas
openai
neo4j>=5.0.0
streamlit